refactor(types): move ResultType to src/types/common/result.ts (correct architectural layer)
Repatriate the canonical `ResultType<T, E>` and `AsyncResult<T, E>` type
definitions from `src/services/core/types/BaseTypes.ts` to a new
`src/types/common/result.ts`. The 9 type-layer importers
(BlkFormat, DataIntegrity, UnitPersistence, UnitSerialization,
UnitTypeHandler, VaultContactTypes, VaultImportExportTypes,
VaultSharingTypes) now import from the new location instead of reaching
back into the service layer.

The `Result` static-factory class (runtime helpers) stays in the service
layer where service code expects it, and BaseTypes re-exports
`ResultType`/`AsyncResult` from the new canonical location for backward
compatibility with all existing service-layer importers.

This eliminates 9 cross-layer import violations where `src/types/`
depended on `src/services/`.
