+++ src/types/unit/UnitTypeHandler.ts
@@ -7,7 +7,7 @@
  * @see openspec/changes/add-multi-unit-type-support/tasks.md Phase 1.3
  */
 
-import type { ResultType } from '@/services/core/types/BaseTypes';
+import type { ResultType } from '@/types/common/result';
 
 import { IBlkDocument } from '../formats/BlkFormat';
 import { UnitCategory } from '../validation/UnitValidationInterfaces';
